requirements.txt
requests==2.32.5
python-dotenv==1.2.1
python-docx==1.2.0
matplotlib==3.10.8
pandas==2.3.3
openpyxl==3.1.2
langchain-community==0.0.30
duckduckgo-search==3.9.10
ddgs==9.10.0
pillow==12.1.0
Pillow==12.1.0
aiohttp==3.13.3
psutil==7.2.1